Fix layout geometry mapping to do correct 3D scene extension.
This fixes MapLocalToAncestor (which is used by getBoundingClientRect),
MapAncestorToLocal, and MapVisualRectToContainer (although users of
these APIs should eventually switch to using GeometryMapper, see
https://crbug.com/1222769) to follow the same rules as
PaintPropertyTreeBuilder (which constructs the state that GeometryMapper
uses) for extension of 3D scenes created by transform-style:
preserve-3d.  These rules change under the TransformInterop feature,
since 3D scene extension should be based on DOM elements rather than
containing blocks, and should not extend across nodes that have no
transform.  However, this patch makes changes both with and without
TransformInterop enabled since the existing code did not match the
rendering behavior.

Testing:
* the getBoundingClientRect calls in
  web_tests/external/wpt/css/css-transforms/3d-rendering-context-behavior.html
  test MapLocalToAncestor, as do the
  MapCoordinatesTest.LocalToAbsoluteTransformFlattens* and
  MapCoordinatesTest.Transform3DWithOffset* unittests.
* The MouseEvent.offsetX getters in various tests in
  web_tests/transforms/3d/point-mapping/ (3d-point-mapping-2.html,
  3d-point-mapping-3.html, 3d-point-mapping-deep.html, and
  3d-point-mapping-preserve-3d.html) test MapAncestorToLocal.

Note that the change to MapVisualRectToContainer is untested.

Note that the new test 3d-point-mapping-2-transforminterop.html is a
variant of 3d-point-mapping-2.html that is intended to pass with
TransformInterop enabled.  However, it does not yet pass due to what I
think is a separate (but probably relatively nearby) bug.

Note also that the changes to map_coordinates_test.cc are needed because
the new non-TransformInterop code leaves the transform in a 3D state
where it didn't before (because there are no elements up to the root
that would break a preserve-3d scene).  While an alternative is trying
to fix this up when we reach the root element or the given ancestor, I'm
not aware of problems it causes beyond the unit test, and it seems like
a reasonable result to me.  Fixing the unit test requires either a call
to matrix.FlattenTo2d() or (as I chose) changing the ProjectPoint calls
to MapPoint.  I don't want to add additional flattening beyond this unit
test without evidence that doing so would make things better rather than
worse.

Bug: 1008483
Change-Id: Ibdf34963152e101ab51655a332eec1d4b0ea014f