Append nested fragmentainers for OOFs if needed.
In a nested fragmentation context (e.g. multicol in multicol), if an OOF
couldn't fit in the fragmentainers than we already had, we'd just put
multiple fragments for the same OOF into the last fragmentainer, with an
additional inline-offset for each fragment.

With an upcoming pre-paint change, this approach is problematic, since
we want to use the fragmentainer index as a unique identifier for each
FragmentData entry.

This CL fixes that. We'll now create as many inner fragmentainers that
we need, and never put more than one fragment for the same node into the
same fragmentainer. This requires us to regenerate inner multicol
container fragments, and overwrite references to the old fragments.

One consequence of fixing this is that we now also get column rules
right in such cases, since column rule painting is based on having
the actual fragmentainers there.

Bug: 1043787, 1079031
Change-Id: Ie94440a199636f5e29f3443d0ae37a8d079c3759