Add base plumbing for sync access handle
This CL adds createSyncAccessHandle() to FileSystemFileHandle, and its
associated scaffolding. When creating a SyncAccessHandle, an OS-level
file descriptor is returned by the browser. This is a temporary state to
enable prototyping. When the file locking implementation becomes
clearer, the browser will return a host object.

Access to this surface is gated behind the
--enable-features=FileSystemAccessAccessHandle content flag.

For more context on this surface and the general objective of the merger
between Storage Foundation and OPFS, check out:
https://docs.google.com/document/d/1g7ZCqZ5NdiU7oqyCpsc2iZ7rRAY1ZXO-9VoG4LfP7fM

Note: WPT tests are not marked as tentative because of
crbug.com/1224358, they will be renamed after the bug is fixed.

Bug: 1218431
Change-Id: I118a00cc6ea1bf2d1643d38f7716d001baa2b5ca